mlx4: Get rid of the mlx4_interface.activate callback
The mlx4_interface.activate callback was introduced in
commit 79857cd31fe7 "net/mlx4: Postpone the registration of net_device".
It dealt with a situation when a netdev notifier received
a NETDEV_REGISTER event for a new net_device created by mlx4_en but the
same device was not yet visible to mlx4_get_protocol_dev(). The callback
can be removed now that mlx4_get_protocol_dev() is gone.

Signed-off-by: Petr Pavlu <petr.pavlu@suse.com>